# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Mate API",
    "version": "1.0.0",
    "description": "Dokumentacja API dla aplikacji Travel Mate - kompleksowej platformy do planowania podróży",
    "contact": {
      "name": "Travel Mate Team",
      "email": "contact@travelmate.com"
    },
    "license": {
      "name": "Private",
      "url": "https://travelmate.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token z Firebase Authentication. Otrzymasz go po zalogowaniu przez POST /api/auth/login"
      }
    },
    "schemas": {
...
      "Accommodation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "properties": {
              "latitude": {
                "type": "number"
              },
              "longitude": {
                "type": "number"
              },
              "address": {
                "type": "string"
              }
            }
          },
          "price": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "number"
              },
              "currency": {
                "type": "string"
              },
              "totalAmount": {
                "type": "number"
              }
            }
          },
          "rating": {
            "type": "number"
          },
          "reviews": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string"
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "link": {
            "type": "string"
          },
          "hotelClass": {
            "type": "number"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [],
...
  "paths": {
    "/api/accommodations/search": {
      "get": {
        "tags": [
          "Accommodations"
        ],
        "summary": "Wyszukaj noclegi",
        "description": "Wyszukuje dostępne noclegi (hotele) za pomocą SerpAPI Google Hotels",
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lokalizacja (miasto lub region)",
            "example": "Ateny, Grecja"
          },
          {
            "in": "query",
            "name": "checkInDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data zameldowania (YYYY-MM-DD)",
            "example": "2026-06-15"
          },
          {
            "in": "query",
            "name": "checkOutDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data wymeldowania (YYYY-MM-DD)",
            "example": "2026-06-22"
          },
          {
            "in": "query",
            "name": "adults",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 2
            },
            "description": "Liczba dorosłych",
            "example": 2
          },
          {
            "in": "query",
            "name": "children",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Liczba dzieci",
            "example": 0
          },
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "type": "string",
              "default": "PLN"
            },
            "description": "Waluta cen",
            "example": "PLN"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista dostępnych noclegów",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "accommodations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Accommodation"
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "accommodations": [
                    {
                      "id": "hotel123",
                      "name": "Hotel Grande Bretagne",
                      "description": "Luksusowy hotel w centrum Aten",
                      "location": {
                        "latitude": 37.9755,
                        "longitude": 23.7348,
                        "address": "Syntagma Square, Athens"
                      },
                      "price": {
                        "amount": 250,
                        "currency": "PLN",
                        "totalAmount": 1750
                      },
                      "rating": 4.7,
                      "reviews": 1250,
                      "imageUrl": "https://example.com/hotel.jpg",
                      "amenities": [
                        "WiFi",
                        "Basen",
                        "Spa"
                      ],
                      "link": "https://booking.com/hotel123",
                      "hotelClass": 5
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe parametry wyszukiwania",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Parametry location, checkInDate i checkOutDate są wymagane"
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera lub SerpAPI",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}